# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "ApplicationsController_getAllJobPostings",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "jobType",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isRemote",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir toutes les offres ouvertes",
        "tags": [
          "Job Applications"
        ]
      },
      "post": {
        "operationId": "ApplicationsController_createJobPosting",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobPostingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer une offre d'emploi",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/search": {
      "get": {
        "operationId": "ApplicationsController_searchJobs",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Rechercher des offres",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{slug}": {
      "get": {
        "operationId": "ApplicationsController_getJobBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Obtenir une offre par slug",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/my/postings": {
      "get": {
        "operationId": "ApplicationsController_getMyJobPostings",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir mes offres postées",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{id}": {
      "put": {
        "operationId": "ApplicationsController_updateJobPosting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobPostingDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour une offre",
        "tags": [
          "Job Applications"
        ]
      },
      "delete": {
        "operationId": "ApplicationsController_deleteJobPosting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer une offre",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{id}/publish": {
      "post": {
        "operationId": "ApplicationsController_publishJobPosting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Publier une offre",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{id}/close": {
      "post": {
        "operationId": "ApplicationsController_closeJobPosting",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Fermer une offre",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{jobId}/apply": {
      "post": {
        "operationId": "ApplicationsController_applyToJob",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Postuler à une offre",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/my/applications": {
      "get": {
        "operationId": "ApplicationsController_getMyApplications",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir mes candidatures",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{jobId}/applications": {
      "get": {
        "operationId": "ApplicationsController_getJobApplications",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir les candidatures d'une offre (recruteur)",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/applications/{id}/status": {
      "put": {
        "operationId": "ApplicationsController_updateApplicationStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour le statut d'une candidature (recruteur)",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/applications/{id}/withdraw": {
      "post": {
        "operationId": "ApplicationsController_withdrawApplication",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Retirer sa candidature",
        "tags": [
          "Job Applications"
        ]
      }
    },
    "/api/jobs/{jobId}/statistics": {
      "get": {
        "operationId": "ApplicationsController_getJobStatistics",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir les statistiques d'une offre (recruteur)",
        "tags": [
          "Job Applications"
        ]
      }
    },
...
  },
  "info": {
    "title": "EduLearn API",
    "description": "API pour la plateforme EduLearn",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateJobPostingDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Développeur Full Stack"
          },
          "description": {
            "type": "string",
            "example": "Nous recherchons un développeur..."
          },
          "requirements": {
            "type": "string",
            "example": "Maîtrise de React, Node.js..."
          },
          "responsibilities": {
            "type": "string"
          },
          "benefits": {
            "type": "string"
          },
          "company": {
            "type": "string",
            "example": "Tech Corp"
          },
          "companyLogo": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "example": "Paris, France"
          },
          "isRemote": {
            "type": "boolean",
            "example": true
          },
          "jobType": {
            "type": "string",
            "enum": [
              "full_time",
              "part_time",
              "contract",
              "internship",
              "freelance"
            ],
            "example": "full_time"
          },
          "experienceLevel": {
            "type": "string",
            "enum": [
              "entry",
              "junior",
              "mid",
              "senior",
              "lead"
            ],
            "example": "mid"
          },
          "salaryMin": {
            "type": "number",
            "example": 40000
          },
          "salaryMax": {
            "type": "number",
            "example": 60000
          },
          "salaryCurrency": {
            "type": "string",
            "example": "EUR"
          },
          "skills": {
            "example": [
              "React",
              "Node.js",
              "TypeScript"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deadline": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "open",
              "closed",
              "archived"
            ],
            "example": "open",
            "default": "open",
            "description": "Statut de l'offre (par défaut: OPEN)"
          }
        },
        "required": [
          "title",
          "description",
          "requirements",
          "company",
          "location",
          "isRemote",
          "jobType",
          "experienceLevel"
        ]
      },
      "UpdateJobPostingDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "open",
              "closed",
              "archived"
            ],
            "description": "Statut de l'offre"
          }
        }
      },
      "CreateApplicationDto": {
        "type": "object",
        "properties": {
          "coverLetter": {
            "type": "string",
            "example": "Je suis très intéressé par ce poste..."
          },
          "resumeUrl": {
            "type": "string",
            "example": "https://example.com/resume.pdf"
          },
          "portfolioUrl": {
            "type": "string",
            "example": "https://portfolio.com"
          },
          "phone": {
            "type": "string",
            "example": "+33612345678"
          },
          "linkedinUrl": {
            "type": "string",
            "example": "https://linkedin.com/in/..."
          },
          "githubUrl": {
            "type": "string",
            "example": "https://github.com/..."
          }
        },
        "required": [
          "coverLetter"
        ]
      },
      "UpdateApplicationStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "reviewing",
              "shortlisted",
              "interview",
              "accepted",
              "rejected",
              "withdrawn"
            ]
          },
          "notes": {
            "type": "string"
          },
          "feedback": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}